Extract related articles' article-reply list dedup mechanism into hooks
